disable check that herb is extremely close to traj start pos
Because OWD published last commanded position, not actual position,
but ros_control and rewd_controllers publish actual position,
the check in herbpy that current position is extremely close to the
planning start position almost always fails.

However, this has not been a problem in practice, and we have been
running with this code commented out for over a month.